ui: Add support for cluster creation from template
